Always use legacy view controller for displaying preview with no siteaccess
eZ 5 only forwarded preview requests to legacy if legacy_mode is enabled for
siteaccess.

Since legacy admin was supposed to always be in legacy mode, it worked fine.
However, when we switched legacy admin not to use legacy mode, the default
siteaccess parameter was always set to be the current siteaccess, which is
a difference on how legacy worked.

Legacy, in case of missing siteaccess parameter, DID NOT use the default
siteaccess to display the preview as eZ 5 does, but it used a special page
that rendered an iframe which then had a siteaccess parameter set.